+++ src/api_v2/session/delete-session.js
