+++ test/api_v2/session/signout.api.test.js
